modularize and format
