+++ services/command-services/user-command-service/rustfmt.toml
@@ -1,0 +1,1 @@
+hard_tabs = true
